[dart2wasm] Remove a duplicated loop in dart:convert
This removes the outer check in `List<int>` element check:

```
if (allowMalformed) {
   ... check code units in `List<int>`, write 0xFF for invalid bytes ...
} else {
   ... check code units in `List<int>`, throw for invalid bytes ...
}
```

and moves the `allowMalformed` check to the inside of the loop, when the
`int` value is negative or more then 255.

`allowMalformed` in UTF8 is used for two things:

1. To allow invalid bytes in `List<int>` argument.
2. To allow invalid UTF-8 encodings.

Both of these will happen rarely, but (1) will almost never happen as
any code that cares about performance will pass a `Uint8List`. So it
makes sense to assume that (1) will hold very rarely, and optimize based
on that.

Change-Id: I3ecec24d29d1cc2ec8f986a87ac3e7d609002d89
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/372060
Reviewed-by: Martin Kustermann <kustermann@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>