[CFE] Fix timeout on crash in dartdoctest
The test uses the incremental compiler, and saves the previous one to
avoid compiling more than it has to. If a compile crashes though, the
incremental compiler will be left in a state where it thinks it's
currently compiling, and will have to wait for the previous compile to
finish. As it had crashed it won't, and we'll thus get a timeout on the
bot.
Here it's fixed by only saving it for next time once we're done.

Change-Id: I20a5ab20438a125487a95147e7a092f2129796a4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/373501
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>