Add code to sort and truncate the candidate suggestions
This code isn't being called anywhere yet, but needs to be available in
order to allow us to convert the handlers over to using
`computeCandidateSuggestions` directly.

I've started wondering whether we want to rename `SuggestionCollector`
to something like `CompletionResult` to reflect the additional
information that it's starting to have and work that it's starting to
do.

Change-Id: Ib989b6649da2228c257fdc0e63f7013769fa2685
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/373323
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Keerti Parthasarathy <keertip@google.com>