Code complete before a method invocation
While the completion scenario is valid, it's probably an unlikely case.
I discovered the issue because it's one of the cases that was frequently
reporting failures in the completion metrics tool. With this change,
the failure count is now down to 249 (1.96% of attempted completions).

Change-Id: I1c9104352c14eb758040e3e0e3a2d2fafc4f909d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/373780
Reviewed-by: Keerti Parthasarathy <keertip@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>