[cfe] Turn wildcard locals into ExpressionStatements or EmptyStatements.
This CL converts VariableDeclarations to ExpressionStatements or EmptyStatements.
If the declaration has an initializer, it will be converted to an ExpressionStatement, otherwise, an EmptyStatement.

This will help the VM and potentially other backends, avoid allocating space for a variable declaration that will never be used.

Bug: https://github.com/dart-lang/sdk/issues/55655
Change-Id: I845b7e5e61fbb0340623ec0dee7269491bc09513
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/373640
Commit-Queue: Kallen Tu <kallentu@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>