[dart2wasm] Fix bug in handling of optional arguments
The upper bounding mechanism when determining signature of a selector
can often end up making parameter types have top type.

If any parameter is optional and the caller doesn't pass the parameter,
the caller will instead pass a sentinel value. This sentinel value is a
dummy instance that corresponds to the wasm type of the parameter.

=> The dummy instance may be an instance of top type but the actual
type of parameter may be `String?` (and we may even know that it
can only have `OneByteString`s in it)

=> We first have to handle the sentinel before possibly downcasting the
value.

Fixes https://github.com/dart-lang/sdk/issues/56129

TEST=tests/language/regression_56129_test.dart

Change-Id: Icb21a94baadc3f5ddf09f5b08b4385782f914844
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/374565
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>