[dart2wasm] Use row displacement table for representing RTT class relationships
The main RTT information we need is whether a subclass is that of
another and if so how to translate type arguments from the subclass to
the super class.

This CL switches the data structure we use to represent this
relationship to a row-displacement table (simlar to how we do method
dispatch).

=> With two array lookups we can determine if a class is a superclass of
another
=> With one more array lookup we obtain the type argument substitution
array.

The packing of the row displacement table is 80-90%.

Issue https://github.com/dart-lang/sdk/issues/55516

Change-Id: I1a8d1205402c2c5e49c44b2f60c32e79a0e429ce
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/374720
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Ömer Ağacan <omersa@google.com>