analyzer: separate uri_does_not_exist from its doc-import counterpart
The first is a compile-time error, and the second is a warning. This
change allows users to ignore one and not the other, both in
source-based `// ignore_for_file` comments, and in analysis options.

This is particularly important for internal use; we want to mark one as
"allowed to ignore" but not the other.

Cq-Include-Trybots: luci.dart.try:flutter-analyze-try,analyzer-win-release-try,pkg-win-release-try
Change-Id: I6c08d18752f2cac6fe0a389532ffb4a0a11a7d96
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/374204
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>