[cfe] Split BuilderFactoryImpl uses of SourceLibraryBuilder
The uses are fully split into the [_parent] used as the parent builder
when creating builders and [_augmentationRoot] which is used when
importing augmentation libraries or adding parts to augmentation libraries.

Both uses of [SourceLibraryBuilder] should be removed once [Builder]s are
replaced with fragments and augmentation libraries are replaced with
parts with their own parts and imports through the enhanced parts feature.

Change-Id: I51366c4db95639dbcc21b52fa9562d4078f9b97b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/375500
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Jens Johansen <jensj@google.com>