[ddc] Add async rewriter to DDC.
To make review easier I've broken out adding the transformation code into a separate change without referencing it yet. A few other changes had to be included with this. Nothing in this CL changes the behavior of DDC yet.

While the rewrite_async.dart file is largely copied from the dart2js version, there are some significant differences:
- The dart2js version uses a "safeName" function hooked into the Namer to assign variables unique names that avoid collisions/shadowing. DDC doesn't have a pre-processing namer like dart2js so instead I made use of TemporaryIds to achieve the same result. TemporaryIds are referenced-based names and they are assigned names by the LocalNamer when the JS is getting printed. This resolves any scoping issues that would come from the variable hoisting the transform does.
- Dart2js doesn't use for..of loops in JS but DDC does. I had to add a lowering for for..of loops that uses the [Symbol.iterator] on the loop iterable and then uses the same lowering strategy as normal for loops.
- Assignment of sourceInformation objects had to change a bit to create a better debugging experience. The source maps still aren't perfect but this is something we can improve on over time.

This also introduces a JS ast Transformer that allows the AST to be modified using the visitor pattern. Since the tree is immutable we create copies of the nodes in the tree. This is slightly expensive but since we only ever process a single member at a time, this should be okay.

Change-Id: I91da1afce4d6649e8177b36ddde098505a24e927
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/374540
Reviewed-by: Nicholas Shahan <nshahan@google.com>
Reviewed-by: Mark Zhou <markzipan@google.com>