[dart2wasm] Refactor code that builds targets for a selector
Instead of storing selector targets as a `Map<int, Reference>` which is
a sparse map, we store it as `List<(Range, Reference)>`.

This takes advantage of the fact that a method is often inherited by
many classes, so instead of checking for individual class ids we can
check for ranges when dispatching.

Future CLs will take advantage of this.

Change-Id: I0ad0ff68e5375524f13c34b5a33562f4db395a7c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/375282
Reviewed-by: Ömer Ağacan <omersa@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>