[cfe] Introduce NameSpace
This adds a NameSpace interface to Scope and uses it for access to
local members. The NameSpace represents entities declared with a library
or class/extension (type) declaration and is to be used for creating
a Scope instead of being the scope itself. As such it is added to
LibraryBuilder and IDeclarationBuilder and used in place of the `scope`
property for accessing these entities.

Change-Id: Id9b03bc8e820ad8f6d185cc95d695ac4a5e59f8f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/376622
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>