An initial sketch of an alternate element model
This is an attempt to see what the element model might look like if we
made the change I outlined earlier. The goal is to see whether the
changes would be positive or negative from an API perspective.

I have removed comments and any overrides for the sake of brevity.

I added a few comments in places where I thought we might want to
discuss additional changes to the current API. If we're going to make a
breaking change it seems reasonable to consider other possible changes.

Change-Id: I41adfc43e4d8781451723be4433dffc36fb3d47a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/368440
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>