[analysis_server] Don't use full-document-format as a fallback when failing to produce minimal edits for range formatting
When formatting over LSP, we run the formatter (which produces an entire new set of formatted code) and compare the unformatted/formatted code to try and produce minimal edits for the format. This allows things like breakpoints to remain stable in the editor (whereas replacing the entire content might cause them to be lost).

The algorithm we use to do this is a very simple diff on the token stream that assumes only whitespace (and commas) will change. If we find another difference, we fall back to just sending the entire document edit.

In the case where the request was to format a _range_ sending back an entire formatted document is never correct, so this changes the handling to return no edits at all if the request was to format a range.

Fixes https://github.com/Dart-Code/Dart-Code/issues/5169 (to not format the entire document) but does not fix the issue that formatting empty collections produces unexpected differences in token streams - that will be fixed in another CL.

Change-Id: Ibaef49a2800d300606f5132e912c9d30b33cb18a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377422
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>