[analysis_server] Prevent the InlayHints handler from locking requests when obtaining a resolved unit
Instead, obtain a resolved unit and then verify the versions still match (and if not, reject the request as stale).

This doesn't eliminate the possibility of a deadlock because the completion handler still does the same - however it's much less likely that completion is running during server initialization (whereas inlay hints for an open file can happen early) and rejecting requests during typing may have more of an impact there (where the user is often typing, and the editor may be compensating for their changes, so a "stale" response may be better than constantly restarting).

If we make changes to scheduling on the server, the ability to pause incoming _requests_ without also pausing incoming _responses_ may provide the best of both worlds here.

Fixes https://github.com/dart-lang/sdk/issues/56311

Change-Id: I34412356da444e8c57f38ebe0d821c805179b0b2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377642
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>