[analysis_server] Handle minimizing formatting edits that remove trailing spaces from comments
When formatting over LSP, we run the formatter (which produces an entire new set of formatted code) and compare the unformatted/formatted code to try and produce minimal edits for the format. This allows things like breakpoints to remain stable in the editor (whereas replacing the entire content might cause them to be lost).

The algorithm we use to do this is a very simple diff on the token stream that assumes only whitespace (and commas) will change. If we find another difference, we fall back to just sending the entire document edit.

Another place this happens is when the formatter changes whitespace within comments (removing trailing spaces). That whitespace is not between tokens, but part of the comments. This change handles that by allowing entire comments to be replaced when the contents is different (stripped of any matching prefix/suffix).

Fixes https://github.com/Dart-Code/Dart-Code/issues/5200

Change-Id: I675f2b00aaa5ea07088d653ec746524f4166598b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377424
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>