Extract shared logic for AST nodes with annotations/comments.
Previously, the following functionality was duplicated among the
various `...FormalParameterImpl` classes and the `AnnotatedNodeImpl`
class (which is the base class for all AST node kinds that have
annotations/comments *other* than formal parameters):

- The implementation fields `_comment` and `_metadata`.
- Constructor logic to initialize those fields.
- The getter/setter pair `documentationComment`.
- The getters `beginToken`, `metadata` and
  `sortedCommentAndAnnotations`.
- The method `_commentIsBeforeAnnotations`.
- The `visitChildren` logic for visiting the comment and annotations
  in source order.

With one exception, this logic is now mostly in a common mixin,
`_AnnotatedNodeMixin`, reducing code duplication.

The one exception is `beginToken`, which can't be shared yet due to
behavioral differences (see
https://github.com/dart-lang/sdk/issues/56313).

Note that the `...FormalParameterImpl` classes do not yet implement
`AnnotatedNode` or `AnnotatedNodeImpl`. I intend to remedy this in a
follow-up CL.

Change-Id: If5eb942e2e1926b6bfafbca94026eb523c47d003
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377240
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>