Make NormalFormalParameter implement AnnotatedNode.
This makes sense since AnnotatedNode is the interface for AST nodes
that can be annotated with either a documentation comment, a list of
annotations (metadata), or both, and NormalFormalParameter matches
that description.

It is not 100% clear why NormalFormalParameter didn't previously
implement AnnotatedNode; my best guess is that when the class
hierarchy for the analyzer AST was first being designed, an effort was
made to make the "implements" hierarchy of the interface classes match
the "extends" hierarchy of the implementation classes, and since
`NormalFormalParameterImpl` extends `FormalParameterImpl` but not
`AnnotatedNodeImpl`, the decision was made for `NormalFormalParameter`
not to extend `AnnotatedNode`.

But there are already plenty of other cases where the "implements"
hierarchy of the interface classes doesn't match the "extends"
hierarchy of the implementation classes, and I'm not aware of this
causing any problems. For example, `AssertInitializer` implements both
`Assertion` and `ConstructorInitializer`, whereas
`AssertInitializerImpl` extends only
`ConstructorInitializerImpl`. Other AST interface classes that
implement multiple super-interfaces include `AssertStatement`,
`AssignmentExpression`, `BinaryExpression`, `CascadeExpression`,
`ConstructorReference`, `DartPattern`, `FunctionExpressionInvocation`,
`FunctionReference`, `Identifier`, `IndexExpression`,
`MethodInvocation`, `PostfixExpression`, `PrefixExpression`,
`PropertyAccess`, `RedirectingConstructorInvocation`,
`RestPatternElement`, `SuperConstructorInvocation`, and `TypeLiteral`.

This paves the way for fixing
https://github.com/dart-lang/sdk/issues/56313, which concerns
inconsistencies in the behavior of `beginToken` between
`NormalFormalParameter` and the other `AnnotatedNode`s.

Change-Id: Id8216efea6ccbd7f02cbc6ad1b8bc8e4405b42fd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377260
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>