Implement LibraryElement2 and LibraryFragment
This adds documentation comments for the related classes.

I also had to rename several members in order to avoid collisions.

I believe that we are now at a point where we could update the tests to
print the library and library fragments (but none of their children),
but I have not included those changes at this point. I'm happy to add
those changes after you've reviewed what's already here, either in this
CL or in a follow-on CL. Please let me know whether you have a preference.

Change-Id: I135e9b998865ac5413ada57f9233dd6d634652d1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377621
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>