[analysis_server] Refactor computing minimal edits from formatting changes
This is a refactor with no functional changes.

The function to compute minimal edits for formatting changes has grown a lot over time and was very long with many nested functions making it difficult to maintain. Additional fixes I'm working on (to handle token changes between consecutive GT tokens and GT_GT or GT_GT_GT) add more to it, so this is an attempt to simplify it by moving it into a class and extracting the nested helper functions to make the logic easier to follow.

I'm pushing this to Gerrit in multiple patch sets in case it's easier to review:

1. Move the function (unchanged) into a class
2. Tidy up the logic slightly to use class variables (so helpers that were using local variables can be extracted later)
3. Move the nested functions (and some private top-level functions) to the class level

Change-Id: I277b458c8e4780ea02629f8cba80db683057ae90
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/377646
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>